Revert "[APPC-3973] Recover wallet from web download"
