+++ app/src/main/java/com/asfoundation/wallet/app_start/AppStart.kt
@@ -2,7 +2,6 @@ package com.asfoundation.wallet.app_start
 
 import com.asfoundation.wallet.di.IoDispatcher
 import com.asfoundation.wallet.onboarding.use_cases.PendingPurchaseFlowUseCase
-import com.asfoundation.wallet.onboarding.use_cases.RestoreGuestWalletUseCase
 import dagger.Reusable
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
@@ -32,10 +31,6 @@ sealed class StartMode {
     val signature: String?
   ) : StartMode()
 
-  data class RestoreGuestWalletFlow(
-    val backup: String,
-  ) : StartMode()
-
   object Subsequent : StartMode()
 }
 
@@ -44,7 +39,6 @@ class AppStartUseCase @Inject constructor(
   private val repository: AppStartRepository,
   private val gpInstallUseCase: GPInstallUseCase,
   private val pendingPurchaseFlowUseCase: PendingPurchaseFlowUseCase,
-  private val restoreGuestWalletUseCase: RestoreGuestWalletUseCase,
   @IoDispatcher ioDispatcher: CoroutineDispatcher
 ) {
   private val scope = CoroutineScope(ioDispatcher)
@@ -60,7 +54,6 @@ class AppStartUseCase @Inject constructor(
 
     val mode = if (firstInstallTime == lastUpdateTime && runs == 1) {
       pendingPurchaseFlowUseCase()
-        ?: restoreGuestWalletUseCase()
         ?: gpInstallUseCase()
         ?: StartMode.Regular
     } else {
