+++ app/src/main/java/com/asfoundation/wallet/main/MainActivity.kt
@@ -84,11 +84,6 @@ class MainActivity : AppCompatActivity(),
       MainActivitySideEffect.NavigateToOnboarding -> navigator.navigateToOnboarding(
         navController
       )
-      is MainActivitySideEffect.NavigateToOnboardingRecoverGuestWallet ->
-        navigator.navigateToOnboardingRecoverGuestWallet(
-          navController,
-          sideEffect.backup
-        )
       MainActivitySideEffect.NavigateToNavigationBar -> navigator.navigateToNavBarFragment(
         navController
       )
