+++ app/src/main/java/com/asfoundation/wallet/main/MainActivityViewModel.kt
@@ -1,13 +1,11 @@
 package com.asfoundation.wallet.main
 
-import android.util.Log
 import androidx.lifecycle.SavedStateHandle
 import com.appcoins.wallet.core.arch.BaseViewModel
 import com.appcoins.wallet.core.utils.android_common.RxSchedulers
 import com.appcoins.wallet.core.arch.SideEffect
 import com.appcoins.wallet.core.arch.ViewState
 import com.asfoundation.wallet.home.usecases.DisplayConversationListOrChatUseCase
-import com.asfoundation.wallet.main.use_cases.GetCachedGuestWalletUseCase
 import com.asfoundation.wallet.main.use_cases.HasAuthenticationPermissionUseCase
 import com.asfoundation.wallet.main.use_cases.IncreaseLaunchCountUseCase
 import com.asfoundation.wallet.onboarding.use_cases.ShouldShowOnboardingUseCase
@@ -22,7 +20,6 @@ sealed class MainActivitySideEffect : SideEffect {
   object NavigateToNavigationBar : MainActivitySideEffect()
   object NavigateToAutoUpdate : MainActivitySideEffect()
   object NavigateToFingerprintAuthentication : MainActivitySideEffect()
-  data class NavigateToOnboardingRecoverGuestWallet(val backup: String) : MainActivitySideEffect()
 }
 
 object MainActivityState : ViewState
@@ -35,7 +32,6 @@ class MainActivityViewModel @Inject constructor(
   private val hasRequiredHardUpdateUseCase: HasRequiredHardUpdateUseCase,
   private val hasAuthenticationPermissionUseCase: HasAuthenticationPermissionUseCase,
   private val shouldShowOnboardingUseCase: ShouldShowOnboardingUseCase,
-  private val getCachedGuestWalletUseCase: GetCachedGuestWalletUseCase,
   private val savedStateHandle: SavedStateHandle,
   private val rxSchedulers: RxSchedulers
 ) : BaseViewModel<MainActivityState, MainActivitySideEffect>(MainActivityState) {
@@ -55,23 +51,8 @@ class MainActivityViewModel @Inject constructor(
           hasAuthenticationPermissionUseCase() && !authComplete -> {
             sendSideEffect { MainActivitySideEffect.NavigateToFingerprintAuthentication }
           }
-          shouldShowOnboardingUseCase() -> {
-            getCachedGuestWalletUseCase()
-              .subscribeOn(rxSchedulers.io)
-              .observeOn(rxSchedulers.main)
-              .doOnSuccess { backup ->
-                if (backup.isNullOrBlank())
-                  sendSideEffect { MainActivitySideEffect.NavigateToOnboarding }
-                else
-                  sendSideEffect {
-                    MainActivitySideEffect.NavigateToOnboardingRecoverGuestWallet(backup)
-                  }
-              }
-              .doOnError {
-                sendSideEffect { MainActivitySideEffect.NavigateToOnboarding }
-              }
-              .scopedSubscribe()
-          }
+          shouldShowOnboardingUseCase() ->
+            sendSideEffect { MainActivitySideEffect.NavigateToOnboarding }
           else ->
             sendSideEffect { MainActivitySideEffect.NavigateToNavigationBar }
         }
