--- app/src/main/java/com/asfoundation/wallet/main/use_cases/GetCachedGuestWalletUseCase.kt
@@ -1,15 +1,0 @@
-package com.asfoundation.wallet.main.use_cases
-
-import com.asfoundation.wallet.repository.CachedGuestWalletRepository
-import io.reactivex.Single
-import javax.inject.Inject
-
-class GetCachedGuestWalletUseCase @Inject constructor(
-  private val cachedGuestWalletRepository: CachedGuestWalletRepository
-) {
-
-  operator fun invoke(): Single<String?> {
-    return cachedGuestWalletRepository.getCachedGuestWallet()
-  }
-
-}
