+++ app/src/main/java/com/asfoundation/wallet/onboarding/OnboardingViewModel.kt
@@ -2,33 +2,17 @@ package com.asfoundation.wallet.onboarding
 
 import android.net.Uri
 import androidx.lifecycle.viewModelScope
-import com.appcoins.wallet.core.analytics.analytics.legacy.WalletsAnalytics
-import com.appcoins.wallet.core.analytics.analytics.legacy.WalletsEventSender
 import com.appcoins.wallet.core.arch.BaseViewModel
 import com.appcoins.wallet.core.arch.SideEffect
 import com.appcoins.wallet.core.arch.ViewState
 import com.appcoins.wallet.core.utils.android_common.RxSchedulers
 import com.appcoins.wallet.core.arch.data.Async
-import com.appcoins.wallet.feature.changecurrency.data.currencies.FiatValue
-import com.appcoins.wallet.feature.walletInfo.data.wallet.usecases.UpdateWalletInfoUseCase
-import com.appcoins.wallet.feature.walletInfo.data.wallet.usecases.UpdateWalletNameUseCase
 import com.asfoundation.wallet.app_start.AppStartUseCase
 import com.asfoundation.wallet.app_start.StartMode
-import com.asfoundation.wallet.entity.WalletKeyStore
-import com.asfoundation.wallet.main.use_cases.DeleteCachedGuestWalletUseCase
-import com.asfoundation.wallet.main.use_cases.GetBonusGuestWalletUseCase
 import com.asfoundation.wallet.my_wallets.create_wallet.CreateWalletUseCase
 import com.asfoundation.wallet.onboarding.use_cases.HasWalletUseCase
 import com.asfoundation.wallet.onboarding.use_cases.SetOnboardingCompletedUseCase
-import com.asfoundation.wallet.recover.result.FailedEntryRecover
-import com.asfoundation.wallet.recover.result.RecoverEntryResult
-import com.asfoundation.wallet.recover.result.SuccessfulEntryRecover
-import com.asfoundation.wallet.recover.use_cases.RecoverEntryPrivateKeyUseCase
-import com.asfoundation.wallet.recover.use_cases.SetDefaultWalletUseCase
-import com.asfoundation.wallet.recover.use_cases.UpdateBackupStateFromRecoverUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.reactivex.Completable
-import io.reactivex.Single
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -38,8 +22,6 @@ sealed class OnboardingSideEffect : SideEffect {
   object NavigateToWalletCreationAnimation : OnboardingSideEffect()
   object NavigateToRecoverWallet : OnboardingSideEffect()
   object NavigateToFinish : OnboardingSideEffect()
-  object ShowLoadingRecover : OnboardingSideEffect()
-  data class UpdateGuestBonus(val bonus: FiatValue): OnboardingSideEffect()
 }
 
 data class OnboardingState(
@@ -53,15 +35,6 @@ class OnboardingViewModel @Inject constructor(
   private val rxSchedulers: RxSchedulers,
   private val setOnboardingCompletedUseCase: SetOnboardingCompletedUseCase,
   private val createWalletUseCase: CreateWalletUseCase,
-  private val recoverEntryPrivateKeyUseCase: RecoverEntryPrivateKeyUseCase,
-  private val setDefaultWalletUseCase: SetDefaultWalletUseCase,
-  private val updateWalletInfoUseCase: UpdateWalletInfoUseCase,
-  private val updateWalletNameUseCase: UpdateWalletNameUseCase,
-  private val updateBackupStateFromRecoverUseCase: UpdateBackupStateFromRecoverUseCase,
-  private val getBonusGuestWalletUseCase: GetBonusGuestWalletUseCase,
-  private val deleteCachedGuestWalletUseCase: DeleteCachedGuestWalletUseCase,
-  private val walletsEventSender: WalletsEventSender,
-  private val onboardingAnalytics: OnboardingAnalytics,
   appStartUseCase: AppStartUseCase
 ) :
   BaseViewModel<OnboardingState, OnboardingSideEffect>(initialState()) {
@@ -71,8 +44,6 @@ class OnboardingViewModel @Inject constructor(
       return OnboardingState()
     }
   }
-
-  private var guestBonus = FiatValue()
 
   init {
     handleLaunchMode(appStartUseCase)
@@ -111,73 +82,4 @@ class OnboardingViewModel @Inject constructor(
   fun handleLinkClick(uri: Uri) {
     sendSideEffect { OnboardingSideEffect.NavigateToLink(uri) }
   }
-
-  fun handleRecoverGuestWalletClick(backup: String?) {
-    sendSideEffect { OnboardingSideEffect.ShowLoadingRecover }
-    recoverEntryPrivateKeyUseCase(WalletKeyStore(null, backup ?: ""))
-      .flatMap { setDefaultWallet(it) }
-      .doOnSuccess { handleRecoverResult(it) }
-      .doOnError {
-        walletsEventSender.sendWalletCompleteRestoreEvent(
-          WalletsAnalytics.STATUS_FAIL,
-          it.message
-        )
-      }
-      .scopedSubscribe()
-  }
-
-  private fun setDefaultWallet(recoverResult: RecoverEntryResult): Single<RecoverEntryResult> =
-    when (recoverResult) {
-      is FailedEntryRecover -> Single.just(recoverResult)
-      is SuccessfulEntryRecover -> setDefaultWalletUseCase(recoverResult.address)
-        .mergeWith(updateWalletInfoUseCase(recoverResult.address))
-        .andThen(Completable.fromAction { setOnboardingCompletedUseCase() })
-        .andThen(updateWalletNameUseCase(recoverResult.address, recoverResult.name))
-        .toSingleDefault(recoverResult)
-    }
-
-  private fun handleRecoverResult(recoverResult: RecoverEntryResult) =
-    when (recoverResult) {
-      is SuccessfulEntryRecover -> {
-        walletsEventSender.sendWalletRestoreEvent(
-          WalletsAnalytics.ACTION_IMPORT,
-          WalletsAnalytics.STATUS_SUCCESS
-        )
-        deleteCachedGuest()
-        onboardingAnalytics.sendRecoverGuestWalletEvent(guestBonus.amount.toString(), guestBonus.currency)
-        sendSideEffect { OnboardingSideEffect.NavigateToFinish }
-      }
-      is FailedEntryRecover.InvalidPassword -> {
-      }
-      else -> {
-        walletsEventSender.sendWalletRestoreEvent(
-          WalletsAnalytics.ACTION_IMPORT,
-          WalletsAnalytics.STATUS_FAIL, recoverResult.toString()
-        )
-      }
-    }
-
-  private fun updateWalletBackupState() {
-    updateBackupStateFromRecoverUseCase().scopedSubscribe()
-  }
-
-  fun getGuestWalletBonus(key: String) {
-    getBonusGuestWalletUseCase(key)
-      .doOnSuccess {
-        guestBonus = it
-        sendSideEffect { OnboardingSideEffect.UpdateGuestBonus(it) }
-      }
-      .doOnError {
-        walletsEventSender.sendWalletCompleteRestoreEvent(
-          WalletsAnalytics.STATUS_FAIL,
-          "getGuestWalletBonus: ${it.message}"
-        )
-      }
-      .scopedSubscribe()
-  }
-
-  private fun deleteCachedGuest() {
-    deleteCachedGuestWalletUseCase().scopedSubscribe()
-  }
-
 }
