+++ app/src/test/java/com/asfoundation/wallet/app_start/AppStartUseCaseTest.kt
@@ -3,7 +3,6 @@ package com.asfoundation.wallet.app_start
 import app.cash.turbine.testIn
 import com.asfoundation.wallet.gherkin.coScenario
 import com.asfoundation.wallet.onboarding.use_cases.PendingPurchaseFlowUseCase
-import com.asfoundation.wallet.onboarding.use_cases.RestoreGuestWalletUseCase
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.test.StandardTestDispatcher
@@ -36,7 +35,6 @@ internal class AppStartUseCaseTest {
       data.givenData.repository,
       data.givenData.GPInstallUseCase,
       data.givenData.PendingPurchaseFlowUseCase,
-      data.givenData.RestoreGuestWalletUseCase,
       StandardTestDispatcher(scope.testScheduler)
     )
     m And "subscribed for modes"
@@ -64,7 +62,6 @@ internal class AppStartUseCaseTest {
       data.givenData.repository,
       data.givenData.GPInstallUseCase,
       data.givenData.PendingPurchaseFlowUseCase,
-      data.givenData.RestoreGuestWalletUseCase,
       StandardTestDispatcher(scope.testScheduler)
     )
 
@@ -94,7 +91,6 @@ internal class AppStartUseCaseTest {
       data.givenData.repository,
       data.givenData.GPInstallUseCase,
       data.givenData.PendingPurchaseFlowUseCase,
-      data.givenData.RestoreGuestWalletUseCase,
       StandardTestDispatcher(scope.testScheduler)
     )
 
@@ -267,7 +263,6 @@ internal class AppStartUseCaseTest {
     val repository: AppStartRepositoryMock = AppStartRepositoryMock(),
     val GPInstallUseCase: GPInstallUseCase = GPInstallUseCaseMock(),
     val PendingPurchaseFlowUseCase: PendingPurchaseFlowUseCase = PendingPurchaseFlowUseCaseMock(),
-    val RestoreGuestWalletUseCase: RestoreGuestWalletUseCase = RestoreGuestWalletUseCaseMock(),
   )
 
   internal data class ThenData(
@@ -317,15 +312,8 @@ class GPInstallUseCaseMock(
 class PendingPurchaseFlowUseCaseMock(
   private val pendingPurchase: StartMode.PendingPurchaseFlow? = null
 ) : PendingPurchaseFlowUseCase {
+
   override operator fun invoke(): StartMode.PendingPurchaseFlow? {
     return pendingPurchase
-  }
-}
-
-class RestoreGuestWalletUseCaseMock(
-  private val restoreGuestWallet: StartMode.RestoreGuestWalletFlow? = null
-) : RestoreGuestWalletUseCase {
-  override operator fun invoke(): StartMode.RestoreGuestWalletFlow? {
-    return restoreGuestWallet
   }
 }
