--- core/network/backend/src/main/java/com/appcoins/wallet/core/network/backend/api/CachedGuestWalletApi.kt
@@ -1,20 +1,0 @@
-package com.appcoins.wallet.core.network.backend.api
-
-import com.appcoins.wallet.core.network.backend.model.CachedGuestWalletResponse
-import io.reactivex.Completable
-import io.reactivex.Single
-import retrofit2.Response
-import retrofit2.http.DELETE
-import retrofit2.http.GET
-import retrofit2.http.Header
-
-interface CachedGuestWalletApi {
-  @GET("appc/guest_wallet/cached_values")
-  suspend fun getCachedGuestWallet(): Response<CachedGuestWalletResponse>
-
-  @DELETE("appc/guest_wallet")
-  fun deleteCachedGuestWallet(
-    @Header("authorization") authorization: String
-  ): Completable
-
-}
