--- core/network/backend/src/main/java/com/appcoins/wallet/core/network/backend/model/CachedGuestWalletResponse.kt
@@ -1,7 +1,0 @@
-package com.appcoins.wallet.core.network.backend.model
-
-import com.google.gson.annotations.SerializedName
-
-data class CachedGuestWalletResponse(
-  @SerializedName("private_key") val privateKey: String?,
-)
